Moved FuncT<T> to FuncT.h
